+++ lib/govuk/dummy_content_store.rb
@@ -6,6 +6,7 @@ module Govuk
     autoload :ExampleContentItem, "govuk/dummy_content_store/example_content_item"
     autoload :ExampleRepository, "govuk/dummy_content_store/example_repository"
     autoload :LiveRepository, "govuk/dummy_content_store/live_repository"
+    autoload :RandomRepository, "govuk/dummy_content_store/random_repository"
     autoload :Index, "govuk/dummy_content_store/index"
   end
 end