+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Add endpoint for random content
+
 # 0.0.5
 
 * Add the fall back to live feature